Suppress a compiler warning in MythPlayer.
git-svn-id: http://svn.mythtv.org/svn/trunk@25867 7dbf422c-18fa-0310-86e9-fd20926502f2

+++ mythtv/libs/libmythtv/mythplayer.cpp
@@ -1414,7 +1414,7 @@ void MythPlayer::SetCaptionsEnabled(bool enable, bool osd_msg)
         return;
     }
     int mode = NextCaptionTrack(kDisplayNone);
-    if (origMode != mode)
+    if (origMode != (uint)mode)
     {
         DisableCaptions(origMode, false);
